Added Spring web endpoint error handler.
